Memory no longer can be access with global functions.  All memory operations are now off of KMemory.  Some CPU parts of the code can look at internal functions, but the rest of the app should only see a much reduced cleaner API in KMemory.h.  I'm really happy to have cleaned up all the #ifdef's that were in the old Memory.h

+++ source/ui/controls/waitDlg.cpp
@@ -37,7 +37,7 @@ void WaitDlg::run() {
 }
 
 void WaitDlg::addSubLabel(BString subLabel, int max) {
-    while (subLabels.size() > max) {
+    while ((int)subLabels.size() > max) {
         subLabels.pop_front();
     }
     this->subLabels.push_back(subLabel);
